refactor(server): extract language discovery into src/server/language-discovery.ts
DEFAULT_LANGUAGES, getDefaultLanguages, ensureLanguage,
filterDisabledLanguages, the LanguageMetadata table and the registry
discovery chain (dynamic listLanguages -> registered factories -> shipped
defaults, with the container-runtime python guarantee) move off the class
verbatim as discoverSupportedLanguages(registry, logger) and
buildLanguageMetadata(languages). The process.env.MCP_CONTAINER reads move
with the code they belong to and are otherwise untouched.

getSupportedLanguagesAsync / getLanguageMetadata stay on DebugMcpServer as
one-line public delegates: they are ToolContext services (the registry is
read through getAdapterRegistry() at call time, so the server-coverage tests
that swap sessionManager after construction still see the swap).
